requirements.txt
mysql-connector==2.2.9
python-dotenv==0.20.0
